requirements.txt
fastapi==0.91.0
python-logging-loki==0.3.1